Addded the fade before death and winning
Added the screen fades before death and before winning.

Also fixed a few bugs.

The application variables IsGameover and hasPlayerWon are the boolens that trigger it. So if you want to use it in other games make sure to set them.

The MaxTimerDuration will change the speed of the fade.